Refactor environment module
